Remove this file from r0.001
Good new.  Github is unable to commit this files under the branch
r000.0001. The error message is commit fail.

This files should not be able to removed from r000.001 because it is
under the r0.001 but not r000.001.  It is interesting to see how the
Github works in this case.